Time: 101 ms (35.02%), Space: 77.1 MB (18.85%) - LeetHub

+++ 2006-find-the-student-that-will-replace-the-chalk/2006-find-the-student-that-will-replace-the-chalk.cpp
@@ -5,9 +5,9 @@ public:
         for(int i = 0; i<chalk.size(); i++){
             sum+= chalk[i];
         }
-        k = (long long)k%sum;
+        k = k%sum;
         for(int i = 0;i<chalk.size(); i++){
-            if((long long)k < chalk[i]){
+            if(k < chalk[i]){
                 return i;
             }
             k -= chalk[i];
